wrapper: It can be continuable
The .sh will not overwrite the downloaded images

+++ set-a-random-wallpaper.sh
@@ -3,23 +3,40 @@ set -e
 
 BASEDIR="$(dirname "$0")"
 OUTDIR="${HOME}/Pictures/wallpapers/random"
+
+LEFT_SUFFIX='_a-left.jpg'
+RIGHT_SUFFIX='_b-right.jpg'
+CENTER_SUFFIX='_c-center.jpg'
 NAME_PREFIX="$(date +%Y-%V)"
+
 OUTFILE_PREFIX="${OUTDIR}/${NAME_PREFIX}"
+DOUBLE_PATH="${OUTFILE_PREFIX}.jpg"
+LEFT_PATH="${OUTFILE_PREFIX}${LEFT_SUFFIX}"
+RIGHT_PATH="${OUTFILE_PREFIX}${RIGHT_SUFFIX}"
+CENTER_PATH="${OUTFILE_PREFIX}${CENTER_SUFFIX}"
+
 
 # Main
-mkdir -p "${OUTDIR}"
+if [ ! -e "${LEFT_PATH}" ] || [ ! -e "${RIGHT_PATH}" ] || [ ! -e "${CENTER_PATH}" ]
+then
+    if [ ! -e "${DOUBLE_PATH}" ]
+    then
+        mkdir -p "${OUTDIR}"
 
-python "${BASEDIR}/download-random-double-wide-wallpaper.py" "${OUTFILE_PREFIX}.jpg" "$@"
-echo
+        python "${BASEDIR}/download-random-double-wide-wallpaper.py" "${DOUBLE_PATH}" "$@"
+        echo
+    fi
 
-python "${BASEDIR}/crop-wallpaper-for-multi-monitors.py" "${OUTFILE_PREFIX}.jpg" "${OUTFILE_PREFIX}"
-rm "${OUTFILE_PREFIX}.jpg"
-ln -sf "${NAME_PREFIX}_a-left.jpg" "${OUTDIR}/last_a-left.jpg"
-ln -sf "${NAME_PREFIX}_b-right.jpg" "${OUTDIR}/last_b-right.jpg"
-ln -sf "${NAME_PREFIX}_c-center.jpg" "${OUTDIR}/last_c-center.jpg"
-echo
+    python "${BASEDIR}/crop-wallpaper-for-multi-monitors.py" "${DOUBLE_PATH}" "${OUTFILE_PREFIX}"
+    rm "${OUTFILE_PREFIX}.jpg"
+    echo
+fi
 
-python "${BASEDIR}/set-multi-xfce4-wallpapers.py" "${OUTFILE_PREFIX}_c-center.jpg" "${OUTFILE_PREFIX}_a-left.jpg" "${OUTFILE_PREFIX}_b-right.jpg"
+ln -sf "${NAME_PREFIX}${LEFT_SUFFIX}" "${OUTDIR}/last${LEFT_SUFFIX}"
+ln -sf "${NAME_PREFIX}${RIGHT_SUFFIX}" "${OUTDIR}/last${RIGHT_SUFFIX}"
+ln -sf "${NAME_PREFIX}${CENTER_SUFFIX}" "${OUTDIR}/last${CENTER_SUFFIX}"
+
+python "${BASEDIR}/set-multi-xfce4-wallpapers.py" "${CENTER_PATH}" "${LEFT_PATH}" "${RIGHT_PATH}"
 xfdesktop --reload
 
 echo -e '\nAll done'
